Move SendTokens to a separate file
