+++ pkg/motor/motor.go
@@ -7,8 +7,6 @@ import (
 	"strings"
 
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
-	"github.com/cosmos/cosmos-sdk/types"
-	bt "github.com/cosmos/cosmos-sdk/x/bank/types"
 	"github.com/mr-tron/base58"
 	"github.com/sonr-io/multi-party-sig/pkg/party"
 	"github.com/sonr-io/sonr/pkg/client"
@@ -18,7 +16,6 @@ import (
 	"github.com/sonr-io/sonr/pkg/did/ssi"
 	"github.com/sonr-io/sonr/pkg/host"
 	dp "github.com/sonr-io/sonr/pkg/motor/x/discover"
-	"github.com/sonr-io/sonr/pkg/tx"
 	"github.com/sonr-io/sonr/third_party/types/common"
 	mt "github.com/sonr-io/sonr/third_party/types/motor"
 )
@@ -231,43 +228,4 @@ func (w *motorNodeImpl) AddCredentialVerificationMethod(id string, cred *did.Cre
 	}
 
 	return nil
-}
-
-func (m *motorNodeImpl) SendTokens(req mt.PaymentRequest) (*mt.PaymentResponse, error) {
-	// Build Message
-	fromAddr, err := types.AccAddressFromBech32(req.GetFrom())
-	if err != nil {
-		return nil, err
-	}
-
-	toAddr, err := types.AccAddressFromBech32(req.GetTo())
-	if err != nil {
-		return nil, err
-	}
-
-	amount := types.NewCoins(types.NewCoin("snr", types.NewInt(req.GetAmount())))
-	msg1 := bt.NewMsgSend(fromAddr, toAddr, amount)
-	txRaw, err := tx.SignTxWithWallet(m.Wallet, "/cosmos.bank.v1beta1.MsgSend", msg1)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := m.Cosmos.BroadcastTx(txRaw)
-	if err != nil {
-		return nil, err
-	}
-
-	cwir := &bt.MsgSendResponse{}
-	if err := client.DecodeTxResponseData(resp.TxResponse.Data, cwir); err != nil {
-		return nil, err
-	}
-
-	bal := m.GetBalance()
-
-	return &mt.PaymentResponse{
-		Code:           int32(resp.TxResponse.Code),
-		Message:        resp.TxResponse.Info,
-		TxHash:         resp.TxResponse.TxHash,
-		UpdatedBalance: int32(bal),
-	}, nil
 }
